% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{agg.word}
\alias{agg.word}
\title{agg.word function to aggregate word information to sentence or passage level}
\usage{
agg.word(
  data,
  agg.level = "sentence",
  person.id = "",
  passage.id = "",
  word.pos.sen = "",
  word.pos.pas = "",
  sen.pos.pas = "",
  start.time = "",
  end.time = "",
  time.scale = "centi",
  score = ""
)
}
\arguments{
\item{data}{= word level data}

\item{agg.level}{indicator for sentence or passage level data, default is sentence.}

\item{person.id}{variable name that indicates the unique person identifier.}

\item{passage.id}{variable name hat indicates 
the unique task identifier. In the ORF assessment context, it is the
passage identifier.}

\item{word.pos.sen}{The column name that represents 
the word's specific position sequence in a sentence.}

\item{word.pos.pas}{The column name that represents 
the word's specific position sequence in a passage.}

\item{sen.pos.pas}{The column name that represents 
the sentence's specific position sequence in a passage.}

\item{start.time}{The column name in the data that represents the start time
that the word has been read.}

\item{end.time}{The column name in the data that represents the end time
that the word has been read.}

\item{time.scale}{indicator for the unit of reading time, default is centi.
the value can be centi (hundredth) or sec (second)}

\item{score}{The column name in the data that represents the word correctly 
has been read. 0 or 1 with 1 means correctly read.}
}
\value{
data frame (sentence or passage level)
}
\description{
agg.word function to aggregate word information to sentence or passage level
}
