% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspam.R, R/zzz.R
\docType{package}
\name{bspam-package}
\alias{bspam}
\alias{bspam-package}
\title{bspam: Speed-Accuracy Psychometric Modeling for Binomial Count Outcome Data}
\description{
The \pkg{bspam} package provides tools for fitting and scoring
speed-accuracy psychometric models for repeatedly measured binomial count
outcome data. The model jointly represents accuracy and completion time
through latent person-level accuracy and speed components.

The package supports model calibration and person scoring using
Monte Carlo EM and Bayesian approaches. It also includes functionality for
oral reading fluency (ORF) applications, including model-based WCPM
estimation, testlet models for sentence-level data, censoring-aware
scoring, descriptive summaries, and interactive visualizations.

Most users will interact with \code{\link{prep}}, \code{\link{fit.model}},
\code{\link{scoring}}, and the associated summary and plotting methods.

Provides functions to fit speed-accuracy psychometric models for binomial count outcome data. The models jointly represent accuracy and response time and can be used for model-based calibration and scoring of oral reading fluency (ORF) assessment data. The methods are based on Potgieter, Kamata and Kara (2017) \url{https://arxiv.org/abs/1705.10446} and Kara, Kamata, Potgieter and Nese (2020) \doi{10.1177/0013164419900208}.
}
\references{
Potgieter, C. J., Kamata, A., & Kara, Y. (2017).
An EM algorithm for estimating an oral reading speed and accuracy model.
\emph{arXiv preprint arXiv:1705.10446}.
\url{https://arxiv.org/abs/1705.10446}

Kara, Y., Kamata, A., Potgieter, C. J., & Nese, J. F. T. (2020).
Estimating model-based oral reading fluency: A Bayesian approach with a
binomial-lognormal joint latent model.
\emph{Educational and Psychological Measurement, 80}(5), 847-869.
\doi{10.1177/0013164419900208}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kamataak.github.io/bspam/}
  \item \url{https://github.com/kamataak/bspam}
  \item Report bugs at \url{https://github.com/kamataak/bspam/issues}
}

}
\author{
Kuo Wang <wangkuo@nifty.com>

Akihito Kamata <akamata@smu.edu>

Cornelis J. Potgieter <c.potgieter@tcu.edu>

Joseph F. T. Nese <jnese@uoregon.edu>

Yusuf Kara <ykara@mail.smu.edu>

Sarunya Somsong <sarunya.ss@hotmail.com>

Xin Qiao <xqiao@usf.edu>

\strong{Maintainer}: Kuo Wang \email{wangkuo@nifty.com} (\href{https://orcid.org/0000-0002-4925-0591}{ORCID})

Authors:
\itemize{
  \item Akihito Kamata \email{akamata@smu.edu} (\href{https://orcid.org/0000-0001-9570-1464}{ORCID}) [contributor]
  \item Cornelis J. Potgieter \email{c.potgieter@tcu.edu} (\href{https://orcid.org/0000-0002-1995-6817}{ORCID})
  \item Joseph F. T. Nese \email{jnese@uoregon.edu} (\href{https://orcid.org/0000-0002-9878-7395}{ORCID})
  \item Yusuf Kara \email{ykara@mail.smu.edu} (\href{https://orcid.org/0000-0003-0691-0630}{ORCID})
  \item Sarunya Somsong \email{sarunya.ss@hotmail.com} (\href{https://orcid.org/0000-0002-4974-7346}{ORCID})
  \item Xin Qiao \email{xqiao@usf.edu} (\href{https://orcid.org/0000-0002-3248-7859}{ORCID})
}

}
\keyword{internal}
