% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_create_lists.R
\name{create_data_list}
\alias{create_data_list}
\title{Scoring.Passage function}
\usage{
create_data_list(
  Count,
  logT10,
  MaxN,
  MaxN_score,
  a,
  b,
  alpha,
  beta,
  a_score,
  b_score,
  alpha_score,
  beta_score,
  sigma,
  rho,
  C
)
}
\arguments{
\item{Count}{A vector of observed count outcomes for the passages used for scoring the person.}

\item{logT10}{A vector of log-transformed reading times per 10 words for the passages used for scoring the person.}

\item{MaxN}{A vector containing the maximum possible counts, such as passage lengths, for the observed passages.}

\item{MaxN_score}{A vector containing the maximum possible counts, such as passage lengths, for the passages to be scored.}

\item{a}{A Model parameters related to the count model for the observed passages.}

\item{b}{A Model parameters related to the count model for the observed passages.}

\item{alpha}{A Model parameters related to the time model for the observed passages.}

\item{beta}{A Model parameters related to the time model for the observed passages.}

\item{a_score}{A Model parameters related to the count model for the passages to be scored.}

\item{b_score}{A Model parameters related to the count model for the passages to be scored.}

\item{alpha_score}{A Model parameters related to the time model for the passages to be scored.}

\item{beta_score}{A Model parameters related to the time model for the passages to be scored.}

\item{sigma}{The standard deviation parameter for the latent speed component.}

\item{rho}{The correlation between the latent accuracy and speed components.}

\item{C}{A vector of censoring indicators for the observed passages, where 0 indicates a fully observed passage and 1 indicates a censored passage.}
}
\value{
A list containing data and model parameters formatted for
    passage-level Stan scoring models.
}
\description{
Copyright (C) 2021-2024 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/


Pre-processing data: Function prepares data for Stan model
Enter a single student's data for preparation

Assume a student has read K >= 2 passages
}
