% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_create_lists.R
\name{create_data_list_sentence}
\alias{create_data_list_sentence}
\title{Call by scoring.sentence.censoring function}
\usage{
create_data_list_sentence(
  Count,
  logT10,
  MaxN,
  MaxN_score,
  Passage,
  Passage_score,
  a,
  b,
  alpha,
  beta,
  a_score,
  b_score,
  alpha_score,
  beta_score,
  gamma1,
  gamma2,
  sigma,
  rho,
  rhoTestlet,
  C
)
}
\arguments{
\item{Count}{A vector of observed counts, such as the number of words
read correctly, for each sentence.}

\item{logT10}{A vector of log-transformed reading times per 10 words
for each sentence.}

\item{MaxN}{A vector containing the maximum possible counts, such as
the number of words, for each observed sentence.}

\item{MaxN_score}{A vector containing the maximum possible counts,
such as the number of words, for each sentence to be scored.}

\item{Passage}{A vector of passage indicators identifying the passage
to which each observed sentence belongs.}

\item{Passage_score}{A vector of passage indicators identifying the
passage to which each sentence to be scored belongs.}

\item{a}{Model parameter related to the count model.}

\item{b}{Model parameter related to the count model.}

\item{alpha}{Model parameter related to the time model.}

\item{beta}{Model parameter related to the time model.}

\item{a_score}{Model parameter related to the count model for sentences or passages to be scored.}

\item{b_score}{Model parameter related to the count model for sentences or passages to be scored.}

\item{alpha_score}{Model parameter related to the time model for sentences or passages to be scored.}

\item{beta_score}{Model parameter related to the time model for sentences or passages to be scored.}

\item{gamma1}{Hyperparameter related to the count testlet effect.}

\item{gamma2}{Hyperparameter related to the time testlet effect.}

\item{sigma}{The standard deviation parameter for the latent speed
component.}

\item{rho}{The correlation between the latent accuracy and speed
components.}

\item{rhoTestlet}{The correlation between the count and time testlet
effects.}

\item{C}{A vector of censoring indicators for the observed sentences,
where 0 indicates a fully observed sentence and 1 indicates a
censored sentence.}
}
\value{
A list containing data and model parameters formatted for
    sentence-level Stan scoring models.
}
\description{
Copyright (C) 2021-2024 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/


Pre-processing data: Function prepares data for Stan model
Enter a single student's data for preparation

Assume a student has read K >= 2 passages
}
