% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{fit.model}
\alias{fit.model}
\title{Estimate the model parameters}
\usage{
fit.model(
  data = NA,
  person.id = "",
  task.id = "",
  sub.task.id = "",
  max.counts = "",
  obs.counts = "",
  time = "",
  k.in = 5,
  reps.in = 50,
  ests.in = NA,
  est = "mcem",
  se = "none",
  verbose = FALSE,
  testlet = FALSE
)
}
\arguments{
\item{data}{A data frame or a data generated by the \code{prep} function. 
If this argument is used with prepared data, 
the next five arguments \code{person.id, task.id, 
max.counts, obs.counts, time} should be skipped.}

\item{person.id}{Quoted variable name in \code{data} that indicates 
the unique person identifier.}

\item{task.id}{Quoted variable name in \code{data} that indicates 
the unique task identifier. In the ORF assessment context, it is the
passage identifier.}

\item{sub.task.id}{Quoted variable name in \code{data} that indicates 
the unique sub task identifier. In the ORF assessment context, it is the
sentence identifier. It is required when testlet is TRUE.}

\item{max.counts}{Quoted variable name in \code{data} that indicates 
the number of attempts in the task. In the ORF assessment context,
it is the number of words in the passage.}

\item{obs.counts}{Quoted variable name in \code{data} that indicates 
the number of successful attempts in each task. In the ORF assessment
context, it is the number of words read correctly for the passage.}

\item{time}{Quoted variable name in \code{data} that indicates 
the time in seconds took to complete the task. In the ORF context, 
it is the time took to complete reading the passage.}

\item{k.in}{Numeric, indicating the number of imputations. 
Default is \code{5}.}

\item{reps.in}{Numeric, indicating the number of Monte-Carlo iterations. 
Default is \code{50}.}

\item{ests.in}{An optional list of numeric vectors, indicating 
initial values of the model parameters. If this argument is not given, 
\code{mom} function will be called to generate the initial values.}

\item{est}{Quoted string, indicating the choice of the estimator. It has
to be one of \code{"mcem", "bayes"}. Default is \code{"mcem"}.}

\item{se}{Quoted string, indicating the choice of the standard errors. 
It has to be one of \code{"none", "analytic", "bootstrap"}. 
Default is \code{"none"}.}

\item{verbose}{Boolean. If \code{TRUE}, the summary will be output.
Default is \code{FALSE}.}

\item{testlet}{Boolean. If \code{TRUE}, the fit.model.testlet will be run.
Default is \code{FALSE}.}
}
\value{
MCEM list, bayes list
}
\description{
This is an interface function to estimate the model parameters 
based on the task-level (i.e., passage-level in ORF assessment 
context) accuracy and speed data by implementing 
the Monte Carlo EM algorithm described in Potgieter et al. (2017) or
fully Bayesian method described in Kara et al. (2020).
}
\details{
Additional details...
}
\note{
More & more additional note...
}
\examples{
\donttest{
# example code
MCEM_run <- fit.model(data = passage2,
                      person.id = "id.student",
                      task.id = "id.passage",
                      max.counts = "numwords.pass",
                      obs.counts = "wrc",
                      time = "sec",
                      k.in = 5,
                      reps.in = 50,
                      est = "mcem")

}
}
\references{
Potgieter, N., Kamata, A., & Kara, Y. (2017). An EM algorithm for 
  estimating an oral reading speed and accuracy model. Manuscript submitted 
  for publication.  
  
  Kara, Y., Kamata, A., Potgieter, C., & Nese, J. F. (2020). Estimating 
  model-based oral reading fluency: A bayesian approach with a 
  binomial-lognormal joint latent model. 
  Educational and Psychological Measurement, 1-25.
}
\seealso{
\code{\link{scoring}} for scoring.
}
