% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{get.cases}
\alias{get.cases}
\title{Create person-by-occasion case identifiers}
\usage{
get.cases(data)
}
\arguments{
\item{data}{A data frame containing \code{person.id} and
\code{occasion} variables.}
}
\value{
Invisibly returns a one-column data frame containing unique case
    identifiers in the variable \code{cases}.
}
\description{
Creates unique case identifiers by combining the \code{person.id} and
\code{occasion} variables in prepared response data. These identifiers
are used by scoring functions to identify person-by-occasion records.
}
