% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{get.perfectcases}
\alias{get.perfectcases}
\title{Returns perfect cases}
\usage{
get.perfectcases(data)
}
\arguments{
\item{data}{A data frame. A wide-format response data 
generated by the \code{prep} function.}
}
\value{
A vector of perfectly accurate cases.
}
\description{
This is a utility function that returns a list of perfect cases for which 
all responses were correct for all assigned tasks.
}
