% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{get_errlist}
\alias{get_errlist}
\title{To get a string of error passages that have no at least two readers}
\usage{
get_errlist(passage)
}
\arguments{
\item{passage}{}
}
\value{
a string of error passages
}
\description{
To get a string of error passages that have no at least two readers
}
