% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{passage.calib.bayes}
\alias{passage.calib.bayes}
\title{Task calibration example output object by Bayes}
\format{
A list of two elements: \code{$pass.param} is a data frame with
150 rows and 10 variables, and \code{$hyper.param} is a data frame with
1 row and 4 variables.

\code{$pass.param}
\describe{
  \item{\code{a}}{parameter controlling binomial success probabilities}
  \item{\code{b}}{parameter controlling binomial success probabilities}
  \item{\code{alpha}}{parameter controlling reading times}
  \item{\code{beta}}{parameter controlling reading times}
  \item{\code{se_a}}{standard error of a}
  \item{\code{se_b}}{standard error of b}
  \item{\code{se_alpha}}{standard error of alpha}
  \item{\code{se_beta}}{standard error of beta}
  \item{\code{passage.id}}{passage ID}
  \item{\code{nwords.p}}{total number of words in the passage}
}

\code{$hyper.param}
\describe{
  \item{\code{vartau}}{variance of latent speed ability tau}
  \item{\code{rho}}{correlation between the two latent variables}
  \item{\code{se_vartau}}{standard error of vartau}
  \item{\code{se_rho}}{standard error of rho}
}
}
\source{
\url{https://jnese.github.io/core-blog/}
}
\usage{
passage.calib.bayes
}
\description{
This is an example calibration output object obtained using
the Bayesian estimator in \code{fit.model}, with \code{est = "bayes"}.
The object contains calibrated oral reading fluency passage parameters
together with population-level hyperparameters.
}
\references{
Nese, J. F. T. & Kamata, A. (2014-2018). Measuring Oral Reading Fluency:
Computerized Oral Reading Evaluation (Project No. R305A140203) [Grant].
Institute of Education Sciences, U.S. Department of Education.
}
\keyword{datasets}
