% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_sentence_censoring.R
\name{scoring.sentence.censoring}
\alias{scoring.sentence.censoring}
\title{scoring.sentence.censoring function}
\usage{
scoring.sentence.censoring(
  Count = NULL,
  logT10 = NULL,
  N = NULL,
  N_score = NULL,
  Passage = NULL,
  Passage_score = NULL,
  a = NULL,
  a_score = NULL,
  b = NULL,
  b_score = NULL,
  alpha = NULL,
  alpha_score = NULL,
  beta = NULL,
  beta_score = NULL,
  gamma1 = NULL,
  gamma2 = NULL,
  sigma = NULL,
  rho = NULL,
  rhoTestlet = NULL,
  C = NULL,
  type = NULL
)
}
\arguments{
\item{Count}{A vector of observed counts, such as the number of words
read correctly, for each sentence.}

\item{logT10}{A vector of log-transformed reading times per 10 words
for each sentence.}

\item{N}{A vector containing the maximum possible counts, such as
the number of words, for each observed sentence.}

\item{N_score}{A vector containing the maximum possible counts,
such as the number of words, for each sentence to be scored.}

\item{Passage}{A vector of passage indicators identifying the passage
to which each observed sentence belongs.}

\item{Passage_score}{A vector of passage indicators identifying the
passage to which each sentence to be scored belongs.}

\item{a}{A Model parameters related to the count model for the
observed sentences.}

\item{a_score}{A Model parameters related to the count model
for the sentences to be scored.}

\item{b}{A Model parameters related to the count model for the
observed sentences.}

\item{b_score}{A Model parameters related to the count model
for the sentences to be scored.}

\item{alpha}{A Model parameters related to the time model for the
observed sentences.}

\item{alpha_score}{A Model parameters related to the time
model for the sentences to be scored.}

\item{beta}{A Model parameters related to the time model for the
observed sentences.}

\item{beta_score}{A Model parameters related to the time
model for the sentences to be scored.}

\item{gamma1}{Hyperparameters related to the testlet effects
in the count and time models, respectively.}

\item{gamma2}{Hyperparameters related to the testlet effects
in the count and time models, respectively.}

\item{sigma}{The standard deviation parameter for the latent speed
component.}

\item{rho}{The correlation between the latent accuracy and speed
components.}

\item{rhoTestlet}{The correlation between the count and time testlet
effects.}

\item{C}{A vector of censoring indicators for the observed sentences,
where 0 indicates a fully observed sentence and 1 indicates a
censored sentence.}

\item{type}{Output type. Use `"general"` for latent accuracy and speed
estimates and `"orf"` to additionally return model-based WCPM estimates.}
}
\value{
A list containing data and model parameters formatted for
    sentence-level Stan scoring models.
}
\description{
Copyright (C) 2021-2024 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/


Pre-processing data: Function prepares data for Stan model
Enter a single student's data for preparation

Assume a student has read K >= 2 passages
}
