% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspam.R, R/zzz.R
\docType{package}
\name{bspam-package}
\alias{bspam}
\alias{bspam-package}
\title{bspam : A package for fitting the speed-accuracy psychometric model for repeatedly measured count outcome data.}
\description{
`bspam` is an R package that contains functions to fit the speed-accuracy 
psychometric model for repeatedly measured count outcome data (Potgieter, Kamata & Kara, 2017; 
Kara, Kamata, Potgieter & Nese, 2020), where the accuracy is modeled by a binomial 
count latent variable model. For example, the use of this modeling technique 
allows model-based calibration and scoring for oral reading fluency (ORF) assessment data.

This is an R package that contains functions to fit the speed-accuracy psychometric model for count outcome data (Potgieter, Kamata & Kara, 2017; Kara, Kamata, Potgieter & Nese, 2020), where the accuracy is modeled by a binomial count latent variable model. For example, the use of this modeling technique allows model-based calibration and scoring for oral reading fluency (ORF) assessment data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kamataak.github.io/bspam/}
  \item \url{https://github.com/kamataak/bspam}
  \item Report bugs at \url{https://github.com/kamataak/bspam/issues}
}

}
\author{
Kuo Wang <wangkuo@nifty.com>

Akihito Kamata <akamata@smu.edu>

Cornelis J. Potgieter <c.potgieter@tcu.edu>

Joseph F. T. Nese <jnese@uoregon.edu>

Yusuf Kara <ykara@mail.smu.edu>

Sarunya Somsong <sarunya.ss@hotmail.com>

Xin Qiao <xqiao@usf.edu>

\strong{Maintainer}: Kuo Wang \email{wangkuo@nifty.com} (\href{https://orcid.org/0000-0002-4925-0591}{ORCID})

Authors:
\itemize{
  \item Akihito Kamata \email{akamata@smu.edu} (\href{https://orcid.org/0000-0001-9570-1464}{ORCID}) [contributor]
  \item Cornelis J. Potgieter \email{c.potgieter@tcu.edu} (\href{https://orcid.org/0000-0002-1995-6817}{ORCID})
  \item Joseph F. T. Nese \email{jnese@uoregon.edu} (\href{https://orcid.org/0000-0002-9878-7395}{ORCID})
  \item Yusuf Kara \email{ykara@mail.smu.edu} (\href{https://orcid.org/0000-0003-0691-0630}{ORCID})
  \item Sarunya Somsong \email{sarunya.ss@hotmail.com} (\href{https://orcid.org/0000-0002-4974-7346}{ORCID})
  \item Xin Qiao \email{xqiao@usf.edu} (\href{https://orcid.org/0000-0002-3248-7859}{ORCID})
}

}
\keyword{internal}
