% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspam.R
\docType{package}
\name{bspam-package}
\alias{bspam-package}
\title{bspam : A package for fitting the speed-accuracy psychometric model for repeatedly measured count outcome data.}
\description{
\strong{Purpose of this package} 

`bspam` is an R package that contains functions to fit the speed-accuracy 
psychometric model for count outcome data (Potgieter, Kamata & Kara, 2017; 
Kara, Kamata, Potgieter & Nese, 2020), where the accuracy is modeled by a binomial 
count latent variable model. For example, the use of this modeling technique 
allows model-based calibration and scoring for oral reading fluency (ORF) assessment data. 

\strong{Design philosophy} 

Write Design philosophy
}
\author{
Akihito Kamata <akamata@smu.edu>

Cornelis J. Potgieter <c.potgieter@tcu.edu>

Joseph F. T. Nese <jnese@uoregon.edu>

Yusuf Kara <ykara@mail.smu.edu>

Sarunya Somsong <sarunya.ss@hotmail.com>

Kuo Wang <wangkuo@nifty.com>
}
