% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspam.R
\docType{package}
\name{bspam-package}
\alias{bspam-package}
\title{bspam : A package for fitting the speed-accuracy psychometric model for count outcome data.}
\description{
\strong{Purpose of this package} 

Write something here 

\strong{Design philosophy} 

Write Design philosophy
}
\author{
Akihito Kamata <akamata@smu.edu>

Cornelis J. Potgieter <c.potgieter@tcu.edu>

Joseph F. T. Nese <jnese@uoregon.edu>

Yusuf Kara <ykara@mail.smu.edu>

Sarunya Somsong <sarunya.ss@hotmail.com>

Kuo Wang <wangkuo@nifty.com>
}
