% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{desc.data}
\alias{desc.data}
\title{desc.data function to describe data information}
\usage{
desc.data(
  data = NULL,
  person.id = NULL,
  task.id = NULL,
  max.counts = NULL,
  obs.counts = NULL,
  time = NULL,
  sub.task.id = NULL,
  desc.level = "sample",
  verbose = T,
  type = "general",
  testlet = F
)
}
\arguments{
\item{data}{A data frame or a data generated by the \code{prep} function.}

\item{person.id}{Quoted variable name in \code{data} that indicates 
the unique person identifier.}

\item{task.id}{Quoted variable name in \code{data} that indicates 
the unique task identifier. In the ORF assessment context, it is the
passage identifier.}

\item{max.counts}{Quoted variable name in \code{data} that indicates 
the number of attempts in the task. In the ORF assessment context,
it is the number of words in the passage.}

\item{obs.counts}{Quoted variable name in \code{data} that indicates 
the number of successful attempts in each task. In the ORF assessment
context, it is the number of words read correctly for the passage.}

\item{time}{Quoted variable name in \code{data} that indicates 
the time in seconds took to complete the task. In the ORF context, 
it is the time took to complete reading the passage.}

\item{sub.task.id}{Quoted variable name in \code{data} that indicates 
the unique sub task identifier. In the ORF assessment context, it is the
sentence identifier. It is required when testlet is TRUE.}

\item{desc.level}{Quoted string that indicates which data summarization. If \code{"sample"} (default),
will show description of data sample. If \code{"person"}, will output the summarization of person data.}

\item{verbose}{Boolean. If \code{TRUE}, the summary will be output.
Default is \code{TRUE}.}

\item{type}{Quoted string, indication of the choice of output. If \code{"general"} (default),
wcpm scores are not reported. If \code{"orf"}, wcpm scores will be reported.}

\item{testlet}{Boolean. If \code{TRUE}, runs with sub task level, otherwise, with task level.
Default is \code{FALSE}.}
}
\value{
data frame (based on sample data or person information)
}
\description{
desc.data function to describe data information
}
