% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{fit.model}
\alias{fit.model}
\title{This is an interface function to estimate the model parameters.}
\usage{
fit.model(
  data = NA,
  person.data = NA,
  person.id = "",
  task.id = "",
  max.counts = "",
  obs.counts = "",
  time = "",
  k.in = 5,
  reps.in = 2,
  ests.in = NA,
  est = "mcem",
  se = "none",
  verbose = FALSE
)
}
\arguments{
\item{data}{A data frame. It has to be a wide-format response data 
generated by the `prep` function. If this argument is used, 
the next 6 arguments `person.data`, `person.id`, `task.id`, 
`max.counts`, `obs.counts`, and `time` should be skipped.}

\item{person.data}{A data frame. It has be a long-format response 
data object. If this this argument is used, the previous argument `data`
should be skipped.}

\item{person.id}{The column name in `person.data` that represents 
the unique person identifier.}

\item{task.id}{The column name in the data that represents the unique passage identifier.}

\item{max.counts}{The column name in the data that represents the number of words in a passage.}

\item{obs.counts}{The column name in the data that represents the words read correctly for each case.}

\item{time}{The column name in the data that represents the time, in seconds, for each case.}

\item{k.in}{- number of imputations, default is 5}

\item{reps.in}{- number of Monte-Carlo iterations, default is 2}

\item{ests.in}{- if not given, mom function will be called and get est.in output}

\item{est}{- estimator keyword, mcem or bayes, default is mcem}

\item{se}{- standard error keyword / c("none","analytic", "bootstrap"), default is none}

\item{verbose}{- boolean, if shows the summary, default is FALSE}
}
\value{
MCEM list, bayes list
}
\description{
This function estimates the model parameters by implementing 
the Monte Carlo EM algorithm described in Potgieter et al. (2017) or
fully Bayesian method described in Kara et al. (2020).
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/


Additional details...
}
\note{
Additional note...
}
\examples{
# example code
MCEM_run <- fit.model(person.data = passage2,
                      person.id = "id.student",
                      task.id = "id.passage",
                      max.counts = "numwords.pass",
                      obs.counts = "wrc",
                      time = "sec",
                      k.in = 5,
                      reps.in = 50,
                      est = "mcem")

}
\seealso{
[scoring()] for the next step
}
