% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{fit.model}
\alias{fit.model}
\title{Estimate the model parameters}
\usage{
fit.model(
  data = NA,
  person.data = NA,
  person.id = "",
  task.id = "",
  max.counts = "",
  obs.counts = "",
  time = "",
  k.in = 5,
  reps.in = 2,
  ests.in = NA,
  est = "mcem",
  se = "none",
  verbose = FALSE
)
}
\arguments{
\item{data}{A data frame. A wide-format response data 
generated by the \code{prep} function. If this argument is used, 
the next 6 arguments \code{person.data, person.id, task.id, 
max.counts, obs.counts, time} should be skipped.}

\item{person.data}{A data frame. A long-format response 
data object. If this this argument is used, the previous argument 
\code{data} should be skipped.}

\item{person.id}{Quoted variable name in \code{person.data} that indicates 
the unique person identifier.}

\item{task.id}{Quoted variable name in \code{person.data} that indicates 
the unique task identifier. In the ORF assessment context, it is the
passage identifier.}

\item{max.counts}{Quoted variable name in \code{person.data} that indicates 
the number of attempts in the task. In the ORF assessment context,
it is the number of words in the passage.}

\item{obs.counts}{Quoted variable name in \code{person.data} that indicates 
the number of successful attempts in each task. In the ORF assessment
context, it is the number of words read correctly for the passage.}

\item{time}{Quoted variable name in \code{person.data} that indicates 
the time in seconds took to complete the task. In the ORF context, 
it is the time took to complete reading the passage.}

\item{k.in}{Numeric, indicating the number of imputations. 
Default is \code{5}.}

\item{reps.in}{Numeric, indicating the number of Monte-Carlo iterations. 
Default is \code{2}.}

\item{ests.in}{An optional list of numeric vectors, indicating 
initial values of the model parameters. If this argument is not given, 
\code{mom} function will be called to generate the initial values.}

\item{est}{Quoted string, indicating the choice of the estimator. It has
to be one of \code{"mcem", "bayes"}. Default is \code{"mcem"}.}

\item{se}{Quoted string, indicating the choice of the standard errors. 
It has to be one of \code{"none", "analytic", "bootstrap"}. 
Default is \code{"none"}.}

\item{verbose}{Boolean. If \code{TRUE}, the summary will be output.
Default is \code{FALSE}.}
}
\value{
MCEM list, bayes list
}
\description{
This is an interface function to estimate the model parameters by implementing 
the Monte Carlo EM algorithm described in Potgieter et al. (2017) or
fully Bayesian method described in Kara et al. (2020).
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/


Additional details...
}
\note{
Additional note...
}
\examples{
# example code
MCEM_run <- fit.model(person.data = passage2,
                      person.id = "id.student",
                      task.id = "id.passage",
                      max.counts = "numwords.pass",
                      obs.counts = "wrc",
                      time = "sec",
                      k.in = 5,
                      reps.in = 50,
                      est = "mcem")

}
\seealso{
[scoring()] for the next step.
}
