% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testlet.R
\name{fit.model.testlet}
\alias{fit.model.testlet}
\title{This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/}
\usage{
fit.model.testlet(
  data = NULL,
  person.id = "",
  sub.task.id = "",
  obs.counts = "",
  time = "",
  task.id = "",
  max.counts = ""
)
}
\arguments{
\item{data}{A data frame. It has the information of student, passage, sentence, obs.counts and time.}

\item{person.id}{each student's id.}

\item{sub.task.id}{each sentence's id.}

\item{obs.counts}{The column name in the data that represents the words read correctly for each sentence}

\item{time}{The column name in the data that represents the reading time for the sentence.}

\item{task.id}{The column name in the data that represents the unique passage identifier.}

\item{max.counts}{The column name in the data that represents the number of words in a sentence.}
}
\value{
list
}
\description{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/
}
\details{
Additional details...
}
\note{
Additional note...
}
\examples{
# example code
fit.model.testlet <- function(data=NULL, person.id="", sub.task.id="",obs.counts="", time="", task.id="", max.counts="")

}
