% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBootstrapSE.R
\name{getBootstrapSE}
\alias{getBootstrapSE}
\title{Get bootstrap SE
case is a single stu_season_id}
\usage{
getBootstrapSE(
  object,
  person.data,
  case = NA,
  perfect.cases,
  zero.cases,
  est = "map",
  kappa = 1,
  bootstrap = 100,
  external = NULL
)
}
\arguments{
\item{object}{- mcem class object}

\item{person.data}{- Individual response data}

\item{case}{- Case number}

\item{perfect.cases}{- Perfect accurate case}

\item{zero.cases}{- Zero accurate case}

\item{est}{- SE type.(MLE, EAP, and MAP.) default MAP}

\item{kappa}{- Default kappa = 1, better be 5}

\item{bootstrap}{- K number of bootstrap, default is 100}

\item{external}{- if not NULL, will use unread task for estimating}
}
\value{
SE dataset
}
\description{
Added MAP function 07/14/2021
Modified a bug of MLE 07/23/2021
Modified EAP 10/28/2021
}
\details{
This function requires the optional package \pkg{MultiGHQuad}.
Install it before using this function.
}
