% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{passage.calib.bayes}
\alias{passage.calib.bayes}
\title{Task calibration example output object by Bayes}
\format{
two lists: $pass.param with 150 rows and $hyper.param with 4 variables

$pass.param
\describe{
  \item{a}{parameter controlling binomial success probabilities}
  \item{b}{parameter controlling binomial success probabilities}
  \item{alpha}{parameter controlling reading times}
  \item{beta}{parameter controlling reading times}
  \item{se_a}{standard error of a}
  \item{se_b}{standard error of b}
  \item{se_alpha}{standard error of alpha}
  \item{se_beta}{standard error of beta}
  \item{passage.id}{passage ID}
  \item{nwords.p}{the total }
}

$hyper.param
\describe{
  \item{vartau}{variance of latent reading ability tau}
  \item{rho}{correlation between two latent variables}
  \item{se_vartau}{standard error of vartar}
  \item{se_rho}{standard error of rho}
}
}
\source{
\url{https://jnese.github.io/core-blog/}
}
\usage{
passage.calib.bayes
}
\description{
This is an example output object from running the 
\code{fit.model.byaes}
function by using the Bayesian estimator option. It is a result of calibrating 
an oral reading fluency data set from Nese and Kamata (2014-2018) with 
xxxx students on 150 passages.
}
\references{
Nese, J. F. T. & Kamata, A. (2014-2018). Measuring Oral Reading Fluency: 
     Computerized Oral Reading Evaluation (Project No. R305A140203) [Grant]. 
     Institute of Education Sciences, U.S. Department of Education.
}
\keyword{datasets}
