% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{passage2}
\alias{passage2}
\title{Passage-level Oral Reading Fluency assessment data set}
\format{
A data frame with 847 rows and 7 variables.
\describe{
  \item{\code{id.student}}{unique student identifier}
  \item{\code{occasion}}{identifier for longitudinal assessment occasions;
  here a triannual assessment administered in the fall, winter, and spring of a school year}
  \item{\code{grade}}{student grade level}
  \item{\code{id.passage}}{unique passage identifier}
  \item{\code{numwords.pass}}{total number of words in the passage}
  \item{\code{wrc}}{words read correct}
  \item{\code{sec}}{seconds to read the passage}
}
}
\source{
\url{https://jnese.github.io/core-blog/}
}
\usage{
passage2
}
\description{
This is an example data set. It is a passage-level Oral Reading Fluency 
assessment data set for 85 students who were assigned to read 2 to 12 passages 
among the same 12 passages.
The data is a small subset of the data collected by Nese and Kamata (2014-2018).
}
\references{
Nese, J. F. T. & Kamata, A. (2014-2018). Measuring Oral Reading Fluency: 
     Computerized Oral Reading Evaluation (Project No. R305A140203) [Grant]. 
     Institute of Education Sciences, U.S. Department of Education. 
     \url{https://ies.ed.gov/funding/grantsearch/details.asp?ID=1492}
}
\keyword{datasets}
