% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot.task}
\alias{plot.task}
\title{Plot function to show graph of fit.model class}
\usage{
\method{plot}{task}(x, task = NULL, parameter, sort = F, ...)
}
\arguments{
\item{x}{A fit.model object}

\item{task}{task ids for plotting}

\item{parameter}{model parameter for plotting, a,b,alpha,beta}

\item{sort}{sorting flag, TRUE or FALSE}

\item{...}{Additional arguments passed to the method.}
}
\description{
Copyright (C) 2021-2024 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/
}
