% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot.wcpm}
\alias{plot.wcpm}
\title{Plot function to show graph of wcpm information from scoring class}
\usage{
\method{plot}{wcpm}(x, person = NULL, show.se = T, show.abline = T, ...)
}
\arguments{
\item{x}{A scoring object}

\item{person}{Person ids for plotting}

\item{show.se}{tandard error bar flag, TRUE (default) for showing or FALSE for no showing}

\item{show.abline}{abline flag, TRUE (default) for showing or FALSE for no showing}

\item{...}{Additional arguments passed to the method.}
}
\description{
Copyright (C) 2021-2024 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/
}
