% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{prep}
\alias{prep}
\title{This file includes utilities of bspam package.}
\usage{
prep(
  data = data,
  person.id = "",
  task.id = "",
  sub.task.id = "",
  occasion = "",
  group = "",
  max.counts = "",
  obs.counts = "",
  time = "",
  cens = "",
  sentence_level = FALSE
)
}
\arguments{
\item{data}{= student response data}

\item{person.id}{Quoted variable name in \code{person.data} that indicates 
the unique person identifier.}

\item{task.id}{Quoted variable name in \code{person.data} that indicates 
the unique task identifier. In the ORF assessment context, it is the
passage identifier.}

\item{sub.task.id}{Quoted variable name in \code{person.data} that indicates 
the unique sub task identifier. In the ORF assessment context, it is the
sentence identifier. It is required when sentence_level is TRUE.}

\item{occasion}{The column name in the data that represents the unique occasion.}

\item{group}{The column name in the data that represents the unique group.}

\item{max.counts}{Quoted variable name in \code{person.data} that indicates 
the number of attempts in the task. In the ORF assessment context,
it is the number of words in the passage.}

\item{obs.counts}{Quoted variable name in \code{person.data} that indicates 
the number of successful attempts in each task. In the ORF assessment
context, it is the number of words read correctly for the passage.}

\item{time}{Quoted variable name in \code{person.data} that indicates 
the time in seconds took to complete the task. In the ORF context, 
it is the time took to complete reading the passage.}

\item{cens}{The column name in the data that represents the censoring indicators 
whether a specific task or sub task was censored (1) or fully observed (0).}

\item{sentence_level}{flag for sentence or passage level data, default is FALSE}
}
\value{
data list (data.long: data frame,
                   data.wide: list of Y, logT10, N, I)
}
\description{
Copyright (C) 2021-2023 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/


prep function prepares input data for fit.model function
}
