% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{run.mcem}
\alias{run.mcem}
\title{Base function for fitting the speed-accuracy model by MCEM}
\usage{
run.mcem(Y, logT10, N, I, k.in = 5, reps.in = 2, ests.in = NA, verbose = FALSE)
}
\arguments{
\item{Y}{Numeric data matrix of accuracy scores in the size of n x I, 
where n is the number of tasks and I is the number of persons. 
Missing data are allowed.}

\item{logT10}{Numeric data matrix of time in the size of n x I matrix, where
n and I are defined above. The time data should be in the scale of 
log10(times). Missing data are allowed.}

\item{N}{Numeric vector of passage lengths in the length of n.}

\item{I}{Numeric, indicating the number of tasks.}

\item{k.in}{Numeric, indicating the number of imputations. Default is 5.}

\item{reps.in}{Numeric, indicating the number of Monte-Carlo iterations. Default is 2.}

\item{ests.in}{An optional list of numeric vectors, indicating initial values of 
the model param- eters. If this argument is not given, \code{mom} function will 
be called to generate the initial values.}

\item{verbose}{Boolean. If TRUE, the summary will be output. Default is FALSE.}
}
\value{
mcem list
}
\description{
This is a base function for fitting the speed-accuracy model to 
estimate the model parameters by the Monte Carlo EM algorithm 
described in Potgieter et al. (2017). This function is used by the
\code{\link{fit.model}} function, which is recommended for the users to use.
}
\details{
If the user is desired to use this function, note that the response
data file needs to be in the wide format, which can be reshaped from
a long-format response data by the \code{\link{prep.wide}} function.
}
\note{
Update Memo:
10/24/2024 For test
04/29/2021 Modified the mcem function
           based on Nelis's updated.
10/28/2021 Modified the mcem output
}
\references{
Potgieter, N., Kamata, A., & Kara, Y. (2017). An EM algorithm for 
  estimating an oral reading speed and accuracy model. Manuscript submitted 
  for publication.
}
\seealso{
\code{\link{prep.wide}} \code{\link{fit.model}}.
}
