% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R, R/Base_org.R
\name{run.scoring}
\alias{run.scoring}
\title{Base function for scoring by likelihood-based approaches}
\usage{
run.scoring(
  object,
  person.data,
  task.data,
  cases,
  perfect.cases,
  zero.cases,
  est = "map",
  lo = -12,
  hi = 12,
  q = 100,
  kappa = 1,
  external = NULL,
  type = NULL
)

run.scoring(
  object,
  person.data,
  task.data,
  cases,
  perfect.cases,
  zero.cases,
  est = "map",
  lo = -12,
  hi = 12,
  q = 100,
  kappa = 1,
  external = NULL,
  type = NULL
)
}
\arguments{
\item{object}{A class object. Output from calibration phase 
by \code{\link{fit.model}} function.}

\item{person.data}{A data frame. A long-format response data object.}

\item{task.data}{A data frame? Estimated task parameter values?}

\item{cases}{A vector of individual id for which scoring is desired. If no information is 
is specified, it will estimate scores for all cases in the \code{person.data}.}

\item{perfect.cases}{A list of perfect cases (Perfect accurate).}

\item{zero.cases}{A list of zero cases (Zero accurate).}

\item{est}{Quoted string, indicating the choice of the estimator. It has to be one of 
\code{"mle"}, \code{"map"}, \code{"eap"}, \code{"bayes"}. Default is \code{"map"}.}

\item{lo}{Numeric, indicating the lower bound of the quadratures. Default is -12.}

\item{hi}{Numeric, indicating the upper bound of the quadratures. Default is 12.}

\item{q}{Numeric, indicating the number of quadratures. Default is 100.}

\item{kappa}{Numeric, indicating ?? Default is 1.}

\item{external}{An optional vector of task ID's in strings. If \code{NULL} (default), 
the wcpm scores are derived with the tasks the individuals were assigned to. 
If not \code{NULL}, wcpm scores are derived with the tasks provided in the vector, rather
than the tasks the individuals were assigned.}

\item{type}{Quoted string, indication of the choice of output. If \code{"general"} (default),
wcpm scores are not reported. If \code{"orf"}, wcpm scores will be reported.}
}
\value{
wcpm list

wcpm list
}
\description{
This is a base function for estimating factor scores by likelihood-based 
approaches described in Potgieter et al. (2024). This function is used by 
the \code{\link{scoring}} function, which is recommended for the users to use.

This is a base function for estimating factor scores by likelihood-based 
approaches described in Potgieter et al. (2024). This function is used by 
the \code{\link{scoring}} function, which is recommended for the users to use.
}
\details{
Update Memo:
04/29/2021 Modified the wcpm function
           based on Nelis's updated.
06/01/2021 Modified based on Nelis's MAP function.
06/01/2021 Modified based on Sarunya's BiEAP function.
06/20/2021 Modified based on Nelis's updated for MLE and EAP
06/21/2021 Modified a bug of MAP function.
07/12/2021 Modified est.eqs function based on Nelis's code.
07/13/2021 Added Map function for bootstrap.
07/30/2021 Modified wcpm function based on Sarunya's update

Update Memo:
04/29/2021 Modified the wcpm function
           based on Nelis's updated.
06/01/2021 Modified based on Nelis's MAP function.
06/01/2021 Modified based on Sarunya's BiEAP function.
06/20/2021 Modified based on Nelis's updated for MLE and EAP
06/21/2021 Modified a bug of MAP function.
07/12/2021 Modified est.eqs function based on Nelis's code.
07/13/2021 Added Map function for bootstrap.
07/30/2021 Modified wcpm function based on Sarunya's update
}
\references{
Qiao, X, Potgieter, N., & Kamata, A. (2023). Likelihood Estimation of 
  Model-based Oral Reading Fluency. Manuscript submitted 
  for publication.

Qiao, X, Potgieter, N., & Kamata, A. (2023). Likelihood Estimation of 
  Model-based Oral Reading Fluency. Manuscript submitted 
  for publication.
}
