% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{scoring}
\alias{scoring}
\title{Estimate factor scores with task-level model}
\usage{
scoring(
  calib.data = NA,
  data = NA,
  person.id = "",
  task.id = "",
  sub.task.id = "",
  occasion = "",
  group = "",
  max.counts = "",
  obs.counts = "",
  time = "",
  cens = "",
  cases = NULL,
  est = "map",
  se = "analytic",
  failsafe = 0,
  bootstrap = 100,
  external = NULL,
  type = "general",
  censoring = FALSE,
  testlet = FALSE
)
}
\arguments{
\item{calib.data}{A class object. Output from calibration phase 
by \code{\link{fit.model}} function}

\item{data}{A data frame. It is necessary when with 
 the next five arguments \code{person.id, task.id, 
max. counts, obs. counts, time}.}

\item{person.id}{Quoted variable name in \code{data} that 
indicates the unique individual identifier.}

\item{task.id}{Quoted variable name in \code{data} that 
represents the unique task identifier. In the ORF assessment
context, it is the passage identifier.}

\item{sub.task.id}{Quoted variable name in \code{data} that indicates 
the unique sub task identifier. In the ORF assessment context, it is the
sentence identifier. It is required when testlet is TRUE.}

\item{occasion}{The column name in the data that represents the unique occasion.}

\item{group}{The column name in the data that represents the unique group.}

\item{max.counts}{Quoted variable name in \code{data} that 
represents the number of attempts in the task. In the ORF assessment
context, it is the number of words in the passage.}

\item{obs.counts}{The column name in the data that represents the words read correctly for each case.}

\item{time}{The column name in the data that represents the time, in seconds, for each case.}

\item{cens}{The column name in the data that represents the censoring indicators 
whether a specific task or sub task was censored (1) or fully observed (0).
This column is necessary when censoring argument is TRUE.}

\item{cases}{A vector of individual id for which scoring is desired. If no information is 
is specified, it will estimate scores for all cases in the \code{data}.}

\item{est}{Quoted string indicating the estimator. It has to be one of
\code{"mle"}, \code{"map"}, \code{"eap"}, or \code{"bayes"}.
Default is \code{"map"}.}

\item{se}{Quoted string indicating the standard error method. It has to be
one of \code{"analytic"} or \code{"bootstrap"}.
Default is \code{"analytic"}.}

\item{failsafe}{Numeric, indicating the number of retries for bootstrap, which can be set between
0 and 50. Default is 0.}

\item{bootstrap}{Numeric, indicating the number of bootstrap iterations. Default is 100.}

\item{external}{An optional vector of task ID's in strings. If \code{NULL} (default), 
the wcpm scores are derived with the tasks the individuals were assigned to. 
If not \code{NULL}, wcpm scores are derived with the tasks provided in the vector, rather
than the tasks the individuals were assigned.}

\item{type}{Quoted string, indication of the choice of output. If \code{"general"} (default),
wcpm scores are not reported. If \code{"orf"}, wcpm scores will be reported.}

\item{censoring}{Boolean. If \code{TRUE}, interface will call task or sub task censoring.
Default is \code{FALSE}.}

\item{testlet}{Boolean. If \code{TRUE}, runs with sub task level, otherwise, with task level.
This argument is necessary when censoring is TRUE. Default is \code{FALSE}.}
}
\value{
scoring list or Bootstrap dataset or censoring list
}
\description{
This is an interface function to estimated factor scores based on the 
task-level (i.e., passage-level in ORF assessment context) accuracy 
and speed data. It implements likelihood-based approaches (MLE, MAP,
or EAP) described in Qiao et al. (under review) or fully Bayesian 
method described in Kara et al. (2020).
}
\details{
Additional details...
}
\note{
More additional note...
}
\examples{
# example code
WCPM_all <- scoring(calib.data=passage.calib.mcem, 
                   data = passage2,
                   person.id = "id.student",
                   occasion = "occasion",
                   group = "grade",
                   task.id = "id.passage",
                   max.counts = "numwords.pass",
                   obs.counts = "wrc",
                   time = "sec",
                   est = "map", 
                   se = "analytic",
                   type="general")

}
\references{
Qiao, X, Potgieter, N., & Kamata, A. (2023). Likelihood Estimation of 
  Model-based Oral Reading Fluency. Manuscript submitted 
  for publication.  
  
  Kara, Y., Kamata, A., Potgieter, C., & Nese, J. F. (2020). Estimating 
  model-based oral reading fluency: A bayesian approach with a 
  binomial-lognormal joint latent model. 
  Educational and Psychological Measurement, 1–25.
}
\seealso{
\code{\link{fit.model}} for model parameter estimation.
}
