% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{scoring}
\alias{scoring}
\title{Estimate factor scores with task-level model}
\usage{
scoring(
  calib.data = NA,
  person.data = NA,
  person.id = "",
  task.id = "",
  occasion = "",
  group = "",
  max.counts = "",
  obs.counts = "",
  time = "",
  cases = NULL,
  est = "map",
  se = "analytic",
  failsafe = 0,
  bootstrap = 100,
  external = NULL,
  type = "general"
)
}
\arguments{
\item{calib.data}{A class object. Output from calibration phase 
by \code{\link{fit.model}} function}

\item{person.data}{A data frame. A long-format response data object.}

\item{person.id}{Quoted variable name in \code{person.data} that 
indicates the unique individual identifier.}

\item{task.id}{Quoted variable name in \code{person.data} that 
represents the unique task identifier. In the ORF assessment
context, it is the passage identifier.}

\item{occasion}{The column name in the data that represents the unique occasion.}

\item{group}{The column name in the data that represents the unique group.}

\item{max.counts}{Quoted variable name in \code{person.data} that 
represents the number of attempts in the task. In the ORF assessment
context, it is the number of words in the passage.}

\item{obs.counts}{The column name in the data that represents the words read correctly for each case.}

\item{time}{The column name in the data that represents the time, in seconds, for each case.}

\item{cases}{- individual id vectors, will directly use task data if no calib.data provided}

\item{est}{- estimator keyword / c("mle", "map", "eap", "bayes"), default is bayes}

\item{se}{- standard error keyword / c("analytic", "bootstrap"), default is analytic}

\item{failsafe}{- retry time for bootstrap / default 0, can set to 5 ~ 50}

\item{external}{- if not NULL, will use not student read passages for estimating}

\item{type}{- output type, "general" and "orf", default "general" only output tau & theta. "orf" will output wcpm}

\item{bootstrp}{- set K number of bootstrap / default 100}
}
\value{
scoring list or Bootstrap dataset
}
\description{
This is an interface function to estimated factor scores based on the 
task-level (i.e., passage-level in ORF assessment context) accuracy 
and speed data. It implements likelihood-based approaches (MLE, MAP,
or EAP) described in Qiao et al. (under review) or fully Bayesian 
method described in Kara et al. (2020).
}
\details{
Additional details...
}
\note{
More & more additional note...
}
\examples{
# example code
WCPM_all <- scoring(calib.data=MCEM_run, 
                   person.data = passage2,
                   person.id = "id.student",
                   occasion = "occasion",
                   group = "grade",
                   task.id = "id.passage",
                   max.counts = "numwords.pass",
                   obs.counts = "wrc",
                   time = "sec",
                   est = "map", 
                   se = "analytic",
                   type="general")

}
\references{
Qiao, X, Potgieter, N., & Kamata, A. (2023). Likelihood Estimation of 
  Model-based Oral Reading Fluency. Manuscript submitted 
  for publication.  
  
  Kara, Y., Kamata, A., Potgieter, C., & Nese, J. F. (2020). Estimating 
  model-based oral reading fluency: A bayesian approach with a 
  binomial-lognormal joint latent model. 
  Educational and Psychological Measurement, 1–25.
}
\seealso{
\code{\link{fit.model}} for model parameter estimation.
}
