% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fit.model.testlet}
\alias{summary.fit.model.testlet}
\title{summary the information of fit.model.testlet class}
\usage{
\method{summary}{fit.model.testlet}(object, digits = 4, ...)
}
\arguments{
\item{object}{= object}

\item{digits}{= print out numeric with specific digits}

\item{...}{= parameter}
}
\value{
printing information
}
\description{
Copyright (C) 2021-2023 The ORF Project Team
}
\details{
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 3 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.
A copy of the GNU General Public License is available at
http://www.gnu.org/licenses/
}
